go2 w/ depth

+++ active_adaptation/learning/modules/vecnorm.py
@@ -74,7 +74,7 @@ class VecNorm(nn.Module):
         return (input_vector - mean) / std
     
     def _update(self, input_vector: torch.Tensor):
-        input_vector = input_vector.reshape(-1, *self.input_shape)
+        input_vector = input_vector.reshape(-1, *self.input_shape).to(torch.float)
         if len(self.reduction_dims):
             # note that `tensor.mean(())` is not what we want
             sum_ = input_vector.mean(dim=self.reduction_dims, keepdim=True)
